add the related companies filter to the activity page
