+++ src/apps/investments/client/projects/constants.js
@@ -48,21 +48,3 @@ export const PROJECT_STATUS_OPTIONS = [
   { label: 'Lost', value: 'lost' },
   { label: 'Dormant', value: 'dormant' },
 ]
-
-export const INCLUDE_RELATED_COMPANIES = [
-  {
-    label: 'Parent companies',
-    value: 'include_parent_companies',
-  },
-  {
-    label: 'Subsidiary companies',
-    value: 'include_subsidiary_companies',
-  },
-]
-
-export const INCLUDE_RELATED_COMPANIES_DISABLED_SUBSIDIARY =
-  INCLUDE_RELATED_COMPANIES.map((option) =>
-    option.value === 'include_subsidiary_companies'
-      ? { ...option, disabled: 'disabled' }
-      : option
-  )
