+++ src/client/components/ActivityFeed/CollectionList/filters.js
@@ -9,6 +9,7 @@ import {
   buildDatesFilter,
   buildInputFieldFilter,
 } from '../../../filters'
+import { INCLUDE_RELATED_COMPANIES } from '../../RoutedRelatedCompaniesCheckboxGroup/constants'
 
 export const buildSelectedFilters = (
   queryParams,
@@ -87,6 +88,14 @@ export const buildSelectedFilters = (
       options: BUSINESS_INTELLIGENCE_OPTION,
       value: queryParams.was_policy_feedback_provided,
       categoryLabel: LABELS.businessIntelligence,
+    }),
+  },
+  includeRelatedCompanies: {
+    queryParam: 'include_related_companies',
+    options: buildOptionsFilter({
+      options: INCLUDE_RELATED_COMPANIES,
+      value: queryParams.include_related_companies,
+      categoryLabel: LABELS.includeRelatedCompanies,
     }),
   },
 })
